+++ nannou_laser/src/lib.rs
@@ -1,8 +1,11 @@
 //! A cross-platform laser DAC detection and streaming API.
 
 pub extern crate ether_dream;
+pub extern crate helios_dac;
 
-pub mod dac;
+pub mod dac_manager;
+pub mod dac_manager_etherdream;
+pub mod dac_manager_helios;
 #[cfg(feature = "ffi")]
 pub mod ffi;
 #[cfg(feature = "ilda-idtf")]
@@ -11,7 +14,8 @@ pub mod point;
 pub mod stream;
 pub mod util;
 
-pub use dac::{DetectDacs, DetectDacsAsync, DetectedDac, DetectedDacCallback, Id as DacId};
+pub use dac_manager::{DacVariant, DetectDacs, DetectedDac, DetectedDacError, Id as DacId, Result};
+pub use dac_manager_etherdream::{DetectEtherDreamDacsAsync, DetectedDacCallback};
 pub use point::{Point, RawPoint};
 pub use stream::frame::Frame;
 pub use stream::frame::Stream as FrameStream;
@@ -42,23 +46,20 @@ impl Api {
         }
     }
 
-    /// An iterator yielding laser DACs available on the system as they are discovered.
+    /// An iterator yielding laser DACs available on the system per supported DAC variant.
     ///
-    /// Currently, the only laser protocol supported is the ether dream protocol. Thus, this
-    /// enumerates ether dream DACs that are discovered on the LAN.
-    ///
-    /// **Note** that the produced iterator will iterate forever and never terminate unless
+    /// **Note** that the DetectDacs::EtherDream iterator will iterate forever and never terminate unless
     /// `set_timeout` is called on the returned `DetectDacs` instance.
-    pub fn detect_dacs(&self) -> io::Result<DetectDacs> {
-        self.inner.detect_dacs()
+    pub fn detect_dacs(&self, variant: DacVariant) -> Result<DetectDacs> {
+        self.inner.detect_dacs(variant)
     }
 
     /// Block and wait until the DAC with the given `Id` is detected.
-    pub fn detect_dac(&self, id: DacId) -> io::Result<DetectedDac> {
-        self.inner.detect_dac(id)
+    pub fn detect_dac(&self, id: DacId, variant: DacVariant) -> Result<DetectedDac> {
+        self.inner.detect_dac(id, variant)
     }
 
-    /// Spawn a thread for DAC detection.
+    /// Spawn a thread for DAC detection. Currently only implemented for the Etherdream DAC
     ///
     /// Calls the given `callback` with broadcasts as they are received.
     ///
@@ -67,7 +68,7 @@ impl Api {
         &self,
         timeout: Option<Duration>,
         callback: F,
-    ) -> io::Result<DetectDacsAsync>
+    ) -> io::Result<DetectEtherDreamDacsAsync>
     where
         F: 'static + DetectedDacCallback + Send,
     {
@@ -122,19 +123,25 @@ impl Api {
             model,
             render,
             stream_error,
+            is_frame: false,
         }
     }
 }
 
 impl Inner {
     /// See the `Api::detect_dacs` docs.
-    pub(crate) fn detect_dacs(&self) -> io::Result<DetectDacs> {
-        dac::detect_dacs()
+    pub(crate) fn detect_dacs(&self, variant: DacVariant) -> Result<DetectDacs> {
+        match variant {
+            DacVariant::DacVariantEtherdream => {
+                dac_manager_etherdream::detect_dacs().map_err(DetectedDacError::from)
+            }
+            DacVariant::DacVariantHelios => dac_manager_helios::detect_dacs(),
+        }
     }
 
     /// Block and wait until the DAC with the given `Id` is detected.
-    pub(crate) fn detect_dac(&self, id: DacId) -> io::Result<DetectedDac> {
-        for res in self.detect_dacs()? {
+    pub(crate) fn detect_dac(&self, id: DacId, variant: DacVariant) -> Result<DetectedDac> {
+        for res in self.detect_dacs(variant)? {
             let dac = res?;
             if dac.id() == id {
                 return Ok(dac);
@@ -148,11 +155,11 @@ impl Inner {
         &self,
         timeout: Option<Duration>,
         callback: F,
-    ) -> io::Result<DetectDacsAsync>
+    ) -> io::Result<DetectEtherDreamDacsAsync>
     where
         F: 'static + DetectedDacCallback + Send,
     {
-        dac::detect_dacs_async(timeout, callback)
+        dac_manager_etherdream::detect_dacs_async(timeout, callback)
     }
 }